requirements.txt
boto3==1.40.55
kubernetes==34.1.0
PyYAML==6.0.3
